pyproject.toml
[tool.poetry]
name = "torch-prj"
version = "0.1.0"
description = ""
authors = ["maly <swampinhabitant@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gym = {extras = ["accept-rom-license"], version = "^0.26.2"}
pygame = "^2.5.2"
jupyter = "^1.0.0"
notebook = "^7.1.0"
torch = "^2.2.0"
matplotlib = "^3.8.3"
jupytext = "^1.16.1"
ipynb-py-convert = "^0.4.6"
nbconvert = "^7.16.1"
pandas = "^2.2.1"
scikit-learn = "^1.4.1.post1"
flappy-bird-gymnasium = "^0.4.0"
icecream = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
